[DX] Move ChannelRepositoryInterface from bundle to component
