+++ spec/Pim/Bundle/CatalogBundle/Manager/ChannelManagerSpec.php
@@ -6,7 +6,7 @@ use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Manager\CompletenessManager;
 use Pim\Component\Catalog\Model\ChannelInterface;
-use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
+use Pim\Component\Catalog\Repository\ChannelRepositoryInterface;
 
 class ChannelManagerSpec extends ObjectBehavior
 {
