+++ spec/Pim/Bundle/CatalogBundle/Validator/Constraints/FamilyRequirementsValidatorSpec.php
@@ -6,7 +6,7 @@ use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\AttributeRequirementInterface;
 use Pim\Component\Catalog\Model\FamilyInterface;
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
+use Pim\Component\Catalog\Repository\ChannelRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Validator\Constraints\FamilyRequirements;
 use Prophecy\Argument;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
