+++ src/Pim/Bundle/EnrichBundle/Connector/Processor/MassEdit/Family/SetAttributeRequirements.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Family;
 use InvalidArgumentException;
 use Pim\Bundle\CatalogBundle\Factory\AttributeRequirementFactory;
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
+use Pim\Component\Catalog\Repository\ChannelRepositoryInterface;
 use Pim\Bundle\EnrichBundle\Connector\Processor\AbstractProcessor;
 use Pim\Component\Connector\Repository\JobConfigurationRepositoryInterface;
