+++ src/Pim/Bundle/EnrichBundle/MassEditAction/Operation/SetAttributeRequirements.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\EnrichBundle\MassEditAction\Operation;
 use Doctrine\Common\Collections\ArrayCollection;
 use Pim\Bundle\CatalogBundle\Factory\AttributeRequirementFactory;
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
+use Pim\Component\Catalog\Repository\ChannelRepositoryInterface;
 use Pim\Component\Catalog\Model\AttributeRequirementInterface;
 
 /**
