+++ src/Pim/Component/Connector/ArrayConverter/Flat/Product/AttributeColumnInfoExtractor.php
@@ -3,7 +3,7 @@
 namespace Pim\Component\Connector\ArrayConverter\Flat\Product;
 
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
+use Pim\Component\Catalog\Repository\ChannelRepositoryInterface;
 use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
 use Pim\Component\Catalog\Model\AttributeInterface;
